refactor: Double (class) to double (primitive)

+++ TaxiRateComparer/src/com/company/BobsCapsTaxiRating.java
@@ -6,9 +6,9 @@ public class BobsCapsTaxiRating implements ITaxiRating{
         return "Bob's Caps";
     }
 
-    public Double calculatePriceFrom(int distance) {
+    public double calculatePriceFrom(int distance) {
         if (distance > 2) {
-            return (double) (3 * distance);
+            return 3 * distance;
         } else {
             return 9.0;
         }
